src/langgraph/requirements.txt
langgraph==0.2.28
langchain==0.3.7
langchain-openai==0.2.7
openai==1.54.0
supabase==2.9.1
psycopg2-binary==2.9.10
vecs==0.4.3
simple-salesforce==1.12.6
hubspot-api-client==9.0.0
stripe==11.1.0
pika==1.3.2
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
